requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
django-decouple==2.1
django-filter==24.2
django-import-export==4.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-tinymce-4
pandas
openpyxl
python-slugify==8.0.4
reportlab==4.4.3
django-nested-admin==4.1.4
weasyprint==66.0
xhtml2pdf==0.2.17
pdfkit==1.0.0
cloudinary==1.44.1 
django-cloudinary-storage==0.3.0
drf-extra-fields==3.7.0